Move debugging utilities to own module
This saves some imports not required in a regular vimiv session.
