+++ vimiv/utils/__init__.py
@@ -6,21 +6,16 @@
 
 """Various utility functions."""
 
-import cProfile
 import functools
 import inspect
 import re
 from abc import ABCMeta
-from contextlib import contextmanager, suppress
-from datetime import datetime
-from functools import wraps
-from pstats import Stats
+from contextlib import suppress
 from typing import Callable, Optional, List, Any, Iterator, Dict
 
 from PyQt5.QtCore import Qt, pyqtSlot, QRunnable, QThreadPool
 from PyQt5.QtGui import QPixmap, QColor, QPainter
 
-from . import log
 from .customtypes import AnyT, FuncT, FuncNoneT, NumberT
 
 # Different location under PyQt < 5.11
@@ -198,7 +193,7 @@ def asyncfunc(pool: QThreadPool = None) -> Callable[[FuncNoneT], FuncNoneT]:
     """Decorator to run function in parallel on a QThreadPool."""
 
     def decorator(function: FuncNoneT) -> FuncNoneT:
-        @wraps(function)
+        @functools.wraps(function)
         def inner(*args: Any, **kwargs: Any) -> None:
             asyncrun(function, *args, pool=pool, **kwargs)
 
@@ -352,44 +347,3 @@ def create_pixmap(
 
 class AbstractQObjectMeta(wrappertype, ABCMeta):
     """Metaclass to allow setting to be an ABC as well as a QObject."""
-
-
-def timed(function: FuncT) -> FuncT:
-    """Decorator to time a function and log evaluation time."""
-
-    @wraps(function)
-    def inner(*args: Any, **kwargs: Any) -> Any:
-        """Wrap decorated function and add timing."""
-        start = datetime.now()
-        return_value = function(*args, **kwargs)
-        elapsed_in_ms = (datetime.now() - start).total_seconds() * 1000
-        log.info("%s: took %.3f ms", function.__qualname__, elapsed_in_ms)
-        return return_value
-
-    # Mypy seems to disapprove the *args, **kwargs, but we just wrap the function
-    return inner  # type: ignore
-
-
-@contextmanager
-def profile(amount: int = 15) -> Iterator[None]:
-    """Contextmanager to profile code secions.
-
-    Starts a cProfile.Profile upon entry, disables it on exit and prints profiling
-    information.
-
-    Usage:
-        with profile(amount=10):
-            # your code to profile here
-            ...
-        # This is no longer profiled
-
-    Args:
-        amount: Number of lines to restrict the output to.
-    """
-    cprofile = cProfile.Profile()
-    cprofile.enable()
-    yield
-    cprofile.disable()
-    stats = Stats(cprofile)
-    stats.sort_stats("cumulative").print_stats(amount)
-    stats.sort_stats("time").print_stats(amount)
